fix(research): dynamically import the PDF preview component so it only loads on the client
- corrections to the research cards
